+++ src/modules/Elsa.MongoDb/Modules/Labels/LabelStore.cs
@@ -8,23 +8,32 @@ using MongoDB.Driver.Linq;
 
 namespace Elsa.MongoDb.Modules.Labels;
 
+/// <summary>
+/// A MongoDB based store for <see cref="Label"/>s.
+/// </summary>
 public class MongoLabelStore : ILabelStore
 {
     private readonly MongoDbStore<Label> _labelMongoDbStore;
     private readonly MongoDbStore<WorkflowDefinitionLabel> _workflowDefinitionLabelMongoDbStore;
 
+    /// <summary>
+    /// Initializes a new instance of the <see cref="MongoLabelStore"/> class.
+    /// </summary>
     public MongoLabelStore(MongoDbStore<Label> labelMongoDbStore, MongoDbStore<WorkflowDefinitionLabel> workflowDefinitionLabelMongoDbStore)
     {
         _labelMongoDbStore = labelMongoDbStore;
         _workflowDefinitionLabelMongoDbStore = workflowDefinitionLabelMongoDbStore;
     }
 
+    /// <inheritdoc />
     public async Task<Label?> FindByIdAsync(string id, CancellationToken cancellationToken = default) => 
         await _labelMongoDbStore.FindAsync(x => x.Id == id, cancellationToken);
 
+    /// <inheritdoc />
     public async Task<IEnumerable<Label>> FindManyByIdAsync(IEnumerable<string> ids, CancellationToken cancellationToken = default) => 
         await _labelMongoDbStore.FindManyAsync(x => ids.ToList().Contains(x.Id), cancellationToken);
-    
+
+    /// <inheritdoc />
     public async Task<Page<Label>> ListAsync(PageArgs? pageArgs = default, CancellationToken cancellationToken = default)
     {
         if (pageArgs?.Page is null || pageArgs.PageSize is null)
@@ -37,20 +46,24 @@ public class MongoLabelStore : ILabelStore
         var documents = (await _labelMongoDbStore.FindManyAsync(query => Paginate(query, pageArgs), cancellationToken)).ToList();
         return Page.Of(documents, count);
     }
-    
+
+    /// <inheritdoc />
     public async Task SaveAsync(Label record, CancellationToken cancellationToken = default) => 
         await _labelMongoDbStore.SaveAsync(record, cancellationToken);
-    
+
+    /// <inheritdoc />
     public async Task SaveManyAsync(IEnumerable<Label> records, CancellationToken cancellationToken = default) => 
         await _labelMongoDbStore.SaveManyAsync(records, cancellationToken);
-    
+
+    /// <inheritdoc />
     public async Task<bool> DeleteAsync(string id, CancellationToken cancellationToken = default)
     {
         await _workflowDefinitionLabelMongoDbStore.DeleteWhereAsync(x => x.LabelId == id, cancellationToken);
         return await _labelMongoDbStore.DeleteWhereAsync(x => x.Id == id, cancellationToken) > 0;
     }
 
-    public async Task<int> DeleteManyAsync(IEnumerable<string> ids, CancellationToken cancellationToken = default)
+    /// <inheritdoc />
+    public async Task<long> DeleteManyAsync(IEnumerable<string> ids, CancellationToken cancellationToken = default)
     {
         var idList = ids.ToList();
         await _workflowDefinitionLabelMongoDbStore.DeleteWhereAsync(x => idList.Contains(x.LabelId), cancellationToken);
